Refactor(test_generic): Make run script building easier to read
This refactors the process of building the run script, instead
of having multiple lines with a very long run command, the command
is set in an ansible fact which can be used multiple different times.

Additionally, the cmd_line fact uses the collapse operator (>) which
allows us to separate arguments by newlines so the entire run command isn't
occupying a single line, and is easier to maintain.  The collapse
operator removes the newlines and replaces them with spaces.

Co-developed-by: Cursor